fix: split executor tests to satisfy backend-size CI
Move Codex worktree workstation and agent tests into dedicated files so
agent_test.go and workstation_executor_test.go stay under the 1000-line limit.
